Rename `examples.yml` to `example-list.yml`
`examples.yml` isn't really a docassemble interview YAML file. But there's no 
metadata in the file explaining this, so our CI will claim (correctly) that it's
an invalid file.

`docassemble.base:example-list.yml` does the same thing, and we have an exclusion
for that (rightfully). This PR changes the name of the YAML file to match docassemble's
file so that we can ignore it. Open to keeping this repo the same and adding to
`DAYAMLChecker`'s list of things to ignore, but I'm not a huge fan of ignore more files,
in order to handle non-standard non-interview files like this in the `questions` dir
(alternatively, this could go in `sources` or something.